refactor: remove SoupStrainer methods into base class
